Interactive & Standalone support for multi-statement requests (#2428)
* - Interactive support for multi-statement requests where later statements depend on earlier statements.

(cherry picked from commit 5d88cce)

+++ ksql-rest-app/src/test/java/io/confluent/ksql/rest/util/QueryCapacityUtilTest.java
@@ -104,7 +104,7 @@ public class QueryCapacityUtilTest {
         ksqlEngine, ksqlConfig, statementStr);
   }
 
-  private void givenActivePersistentQueries(final long numQueries) {
+  private void givenActivePersistentQueries(final int numQueries) {
     when(ksqlEngine.numberOfPersistentQueries())
         .thenReturn(numQueries);
   }
